Added XML parser for key mappings, test framework.
The Keymap constructor can now accept the filename of a valid XML keymap. The
RapidXML library will parse the XML, and the NameMap classes can translate the
individual strings into keycodes/axisIDs/actions.

Added a test framework to assist with unit testing, and a new test of the Keymap
class, to test out the testing framework, dawg.