Added 'Armed' entity module.
Armed entities maintain a list of attached weapons, which can be fired as a
result of an action. To kick off the arms race, I've added a Missile Entity
(with mesh), and a MissileLauncher weapon to instantiate the Missiles.

I've also touched up the documentation, and fixed a memory leak in the XML
parser.